+++ src/payments/hobex-hps/index.ts
@@ -121,6 +121,7 @@ export {
   isApproved,
   isCanceled,
   isConclusive,
+  isConclusiveAsStatus,
   isInProgress,
   isNoStatement,
   isNotAbortable,